backend/requirements.txt
flask
flask-cors
pillow
pytesseract
sentence-transformers
chromadb
opencv-python
